Split rules engine utils to separate files.
